+++ src/main/kotlin/dev/digamma/shard/ShardBundle.kt
@@ -4,9 +4,9 @@ import com.intellij.DynamicBundle
 import org.jetbrains.annotations.Nls
 import org.jetbrains.annotations.PropertyKey
 
-private const val BUNDLE = "messages.ShardBundle"
-
 object ShardBundle {
+    private const val BUNDLE = "messages.ShardBundle"
+
     private val instance = DynamicBundle(ShardBundle::class.java, BUNDLE)
 
     typealias Key = @PropertyKey(resourceBundle = BUNDLE) String