+++ src/main/kotlin/dev/digamma/shard/ShardFocusManager.kt
@@ -1,8 +1,11 @@
 package dev.digamma.shard
 
 import com.intellij.ide.AppLifecycleListener
+import com.intellij.openapi.application.WriteIntentReadAction
 import com.intellij.openapi.fileEditor.impl.EditorWindowHolder
+import com.intellij.ui.tabs.impl.JBEditorTabs
 import dev.digamma.shard.ex.hierarchy
+import dev.digamma.shard.ex.targetLabel
 import java.awt.Component
 import java.awt.KeyboardFocusManager
 import java.util.*
@@ -16,14 +19,22 @@ object ShardFocusManager {
         lastFocusTime[component] = System.currentTimeMillis()
     }
 
+    private fun processComponent(component: Component) {
+        if (component is EditorWindowHolder) trackFocus(component.editorWindow)
+        if (component is JBEditorTabs) {
+            WriteIntentReadAction.run {
+                if (component.targetLabel?.updateTabActions() == true) {
+                    component.revalidate()
+                    component.repaint()
+                }
+            }
+        }
+    }
+
     object StartupActivity : AppLifecycleListener {
         override fun appFrameCreated(commandLineArgs: List<String?>) {
-            KeyboardFocusManager.getCurrentKeyboardFocusManager().addPropertyChangeListener { event ->
-                (event.newValue as? Component)?.hierarchy?.forEach {
-                    when (it) {
-                        is EditorWindowHolder -> trackFocus(it.editorWindow)
-                    }
-                }
+            KeyboardFocusManager.getCurrentKeyboardFocusManager().addPropertyChangeListener {
+                (it.newValue as? Component)?.hierarchy?.forEach(::processComponent)
             }
         }
     }