+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.1.0
+
+### Added
+
+- Setting to prevent pinned tabs from being closed through the default close action.
+
+### Fixed
+
+- Disable the split action if the active editor cannot be split.
+
 ## 1.0.1
 
 ### Changed